Problem Cause When the extension is loaded, it applies the highlight to the current state of the DOM. However, if the web page updates its content dynamically (for example, through AJAX calls, content loading, or re-rendering), the modifications made by the extension (the highlighting) can be overwritten or removed. As a result, the words do not remain highlighted.
Solution
To solve this problem, we need to make the extension observe changes in the DOM and reapply the highlighting whenever the content changes. This can be done using a MutationObserver in the content script. A MutationObserver observes changes in the DOM and triggers a callback function when such changes occur.

Signed-off-by: Tiago Maluta <maluta@duck.com>